refactor: unify local token history contract
