+++ rust/src/cli/serve/data.rs
@@ -67,7 +67,11 @@ pub async fn cost_response(provider: Option<&str>) -> String {
     for provider_id in selection.as_list() {
         if provider_id == ProviderId::Antigravity {
             let history = crate::providers::antigravity::local_sessions::summarize(30);
-            results.push(antigravity_cost_payload(history, 30));
+            results.push(crate::spend_contract::local_token_history_json(
+                "antigravity",
+                history,
+                30,
+            ));
             continue;
         }
         let (supported, summary) = match provider_id {
@@ -111,31 +115,6 @@ pub async fn cost_response(provider: Option<&str>) -> String {
     json_response(200, serde_json::Value::Array(results))
 }
 
-fn antigravity_cost_payload(
-    history: crate::providers::antigravity::local_sessions::LocalSessionSummary,
-    days: u32,
-) -> serde_json::Value {
-    use crate::providers::antigravity::local_sessions::LocalHistoryCoverage;
-    let coverage = match history.coverage {
-        LocalHistoryCoverage::Complete => "complete",
-        LocalHistoryCoverage::Partial => "partial",
-        LocalHistoryCoverage::Unavailable => "unavailable",
-    };
-    let complete = matches!(history.coverage, LocalHistoryCoverage::Complete);
-    json!({
-        "provider": "antigravity",
-        "supported": true,
-        "days_scanned": days,
-        "cost": {"total_usd": serde_json::Value::Null, "currency": serde_json::Value::Null},
-        "daily": [],
-        "tokens": {"total": complete.then_some(history.total_tokens)},
-        "sessions_count": complete.then_some(history.session_count),
-        "historyCoverage": coverage,
-        "knownZero": complete && history.total_tokens == 0,
-        "note": "Local token history; dollar costs unavailable"
-    })
-}
-
 /// Dashboard-charts shape for one provider's daily spend: [{date, totalCost}].
 fn daily_json(daily: Vec<(String, Option<f64>)>) -> serde_json::Value {
     serde_json::Value::Array(
@@ -169,8 +148,9 @@ mod tests {
         use crate::providers::antigravity::local_sessions::{
             LocalHistoryCoverage, LocalSessionSummary,
         };
-        let complete = antigravity_cost_payload(
-            LocalSessionSummary {
+        let complete = crate::spend_contract::local_token_history_json(
+            "antigravity",
+            LocalTokenHistorySummary {
                 total_tokens: 42,
                 session_count: 1,
                 coverage: LocalHistoryCoverage::Complete,
@@ -181,8 +161,9 @@ mod tests {
         assert_eq!(complete["tokens"]["total"], 42);
         assert_eq!(complete["historyCoverage"], "complete");
 
-        let partial = antigravity_cost_payload(
-            LocalSessionSummary {
+        let partial = crate::spend_contract::local_token_history_json(
+            "antigravity",
+            LocalTokenHistorySummary {
                 total_tokens: 42,
                 session_count: 1,
                 coverage: LocalHistoryCoverage::Partial,
